+++ _invalid_tests/test_wrong.py
@@ -1,0 +1,3 @@
+from nip import load
+
+print(load("../tests/base_tests/raises/configs/wrong_dict_indent.nip"))
