--- tests/_invalid_tests/test_wrong.py
@@ -1,3 +1,0 @@
-from nip import load
-
-print(load("../base_tests/raises/configs/wrong_dict_indent.nip"))
