--- tests/base_tests/comments/builders.py
@@ -1,7 +1,0 @@
-from nip import nip
-
-
-@nip
-class SimpleClass:
-    def __init__(self, name):
-        self.name = name
